feat(connector): integrate Shadow Clipper bridge
