+++ apps/web/src/lib/locales/en.json
@@ -5565,6 +5565,34 @@
     "exportLogsAction": "Export",
     "exportingLogs": "Exporting...",
     "connectorTitle": "This Computer",
+    "clipperTitle": "Shadow Clipper",
+    "clipperDescription": "Connect the browser extension, your library, and local assistants.",
+    "clipperService": "Local connection",
+    "clipperReady": "Running",
+    "clipperStopped": "Stopped",
+    "clipperBrowser": "Browser extension",
+    "clipperBrowserCount": "{{count}} connected",
+    "clipperLibrary": "Local library",
+    "clipperFileCount": "{{count}} files",
+    "clipperLastSync": "Last sync",
+    "clipperNeverSynced": "Not synced yet",
+    "clipperInstallTitle": "Prepare the browser extension",
+    "clipperInstallDescription": "The desktop app finds the local extension and opens Chrome Extensions. Confirm Load unpacked once when Chrome asks.",
+    "clipperPrepareInstall": "Prepare install",
+    "clipperConnectTitle": "Connect the extension",
+    "clipperConnectDescription": "An extension loaded through Prepare install uses the address and token below automatically. Existing installs can enter them manually.",
+    "clipperStart": "Start connection",
+    "clipperStop": "Stop connection",
+    "clipperAddress": "Connection address",
+    "clipperToken": "Connection token",
+    "clipperStartFirst": "Start the connection first",
+    "clipperLoginTitle": "Sync community sign-in",
+    "clipperLoginDescription": "After the extension connects, the desktop app syncs your current community sign-in automatically. You can also sync it again here. Each authorization is one-time and only available to a connected local extension.",
+    "clipperSyncLogin": "Sync current sign-in",
+    "clipperConnectionStarted": "The local connection is running.",
+    "clipperConnectionStopped": "The local connection has stopped.",
+    "clipperInstallPrepared": "The extension folder and Chrome Extensions are open. Choose Load unpacked.",
+    "clipperLoginQueued": "Sign-in sync was sent and will finish on the extension's next connection.",
     "connectorWorkspaceOnlineTitle": "This computer is connected to Shadow",
     "connectorWorkspaceOfflineTitle": "Connect this computer to Shadow",
     "connectorWorkspaceDesc": "Turn local coding tools into community Buddies. Shadow routes conversations and tasks here while your files and runtime stay on this computer.",
