+++ apps/web/src/lib/locales/ja.json
@@ -5225,6 +5225,34 @@
     "exportLogsAction": "書き出す",
     "exportingLogs": "書き出し中...",
     "connectorTitle": "このコンピューター",
+    "clipperTitle": "Shadow Clipper",
+    "clipperDescription": "ブラウザー拡張機能、ライブラリ、ローカルアシスタントを接続します。",
+    "clipperService": "ローカル接続",
+    "clipperReady": "実行中",
+    "clipperStopped": "停止中",
+    "clipperBrowser": "ブラウザー拡張機能",
+    "clipperBrowserCount": "{{count}} 件接続",
+    "clipperLibrary": "ローカルライブラリ",
+    "clipperFileCount": "{{count}} ファイル",
+    "clipperLastSync": "最終同期",
+    "clipperNeverSynced": "未同期",
+    "clipperInstallTitle": "ブラウザー拡張機能を準備",
+    "clipperInstallDescription": "デスクトップアプリがローカル拡張機能と Chrome の拡張機能ページを開きます。Chrome で一度だけ「パッケージ化されていない拡張機能を読み込む」を確認してください。",
+    "clipperPrepareInstall": "インストールを準備",
+    "clipperConnectTitle": "拡張機能を接続",
+    "clipperConnectDescription": "インストールを準備して読み込んだ拡張機能は、下のアドレスとトークンを自動的に使用します。既存の拡張機能には手動で入力できます。",
+    "clipperStart": "接続を開始",
+    "clipperStop": "接続を停止",
+    "clipperAddress": "接続アドレス",
+    "clipperToken": "接続トークン",
+    "clipperStartFirst": "先に接続を開始してください",
+    "clipperLoginTitle": "コミュニティのログインを同期",
+    "clipperLoginDescription": "拡張機能の接続後、現在のコミュニティログインは自動的に同期されます。ここから再同期することもできます。各認証は一度だけ使用され、接続済みのローカル拡張機能にのみ渡されます。",
+    "clipperSyncLogin": "現在のログインを同期",
+    "clipperConnectionStarted": "ローカル接続を開始しました。",
+    "clipperConnectionStopped": "ローカル接続を停止しました。",
+    "clipperInstallPrepared": "拡張機能フォルダーと Chrome の拡張機能ページを開きました。パッケージ化されていない拡張機能を読み込んでください。",
+    "clipperLoginQueued": "ログイン同期を送信しました。拡張機能の次回接続時に完了します。",
     "connectorWorkspaceOnlineTitle": "このコンピューターは Shadow に接続されています",
     "connectorWorkspaceOfflineTitle": "このコンピューターを Shadow に接続",
     "connectorWorkspaceDesc": "ローカルのコーディングツールをコミュニティ Buddy として利用できます。会話とタスクは Shadow がここへ届け、ファイルとランタイムはこのコンピューターに残ります。",
