+++ apps/web/src/lib/locales/ko.json
@@ -5225,6 +5225,34 @@
     "exportLogsAction": "내보내기",
     "exportingLogs": "내보내는 중...",
     "connectorTitle": "이 컴퓨터",
+    "clipperTitle": "Shadow Clipper",
+    "clipperDescription": "브라우저 확장 프로그램, 라이브러리, 로컬 도우미를 연결합니다.",
+    "clipperService": "로컬 연결",
+    "clipperReady": "실행 중",
+    "clipperStopped": "중지됨",
+    "clipperBrowser": "브라우저 확장 프로그램",
+    "clipperBrowserCount": "{{count}}개 연결됨",
+    "clipperLibrary": "로컬 라이브러리",
+    "clipperFileCount": "파일 {{count}}개",
+    "clipperLastSync": "마지막 동기화",
+    "clipperNeverSynced": "아직 동기화되지 않음",
+    "clipperInstallTitle": "브라우저 확장 프로그램 준비",
+    "clipperInstallDescription": "데스크톱 앱이 로컬 확장 프로그램과 Chrome 확장 프로그램 페이지를 엽니다. Chrome에서 압축해제된 확장 프로그램 로드를 한 번 확인하세요.",
+    "clipperPrepareInstall": "설치 준비",
+    "clipperConnectTitle": "확장 프로그램 연결",
+    "clipperConnectDescription": "설치 준비를 통해 로드한 확장 프로그램은 아래 주소와 토큰을 자동으로 사용합니다. 기존 설치에는 직접 입력할 수 있습니다.",
+    "clipperStart": "연결 시작",
+    "clipperStop": "연결 중지",
+    "clipperAddress": "연결 주소",
+    "clipperToken": "연결 토큰",
+    "clipperStartFirst": "먼저 연결을 시작하세요",
+    "clipperLoginTitle": "커뮤니티 로그인 동기화",
+    "clipperLoginDescription": "확장 프로그램이 연결되면 현재 커뮤니티 로그인이 자동으로 동기화됩니다. 여기에서 다시 동기화할 수도 있습니다. 각 인증은 한 번만 사용되며 연결된 로컬 확장 프로그램에만 전달됩니다.",
+    "clipperSyncLogin": "현재 로그인 동기화",
+    "clipperConnectionStarted": "로컬 연결이 시작되었습니다.",
+    "clipperConnectionStopped": "로컬 연결이 중지되었습니다.",
+    "clipperInstallPrepared": "확장 프로그램 폴더와 Chrome 확장 프로그램 페이지가 열렸습니다. 압축해제된 확장 프로그램 로드를 선택하세요.",
+    "clipperLoginQueued": "로그인 동기화가 전송되었으며 확장 프로그램의 다음 연결에서 완료됩니다.",
     "connectorWorkspaceOnlineTitle": "이 컴퓨터가 Shadow에 연결되었습니다",
     "connectorWorkspaceOfflineTitle": "이 컴퓨터를 Shadow에 연결",
     "connectorWorkspaceDesc": "로컬 코딩 도구를 커뮤니티 Buddy로 사용하세요. 대화와 작업은 Shadow가 이곳으로 전달하고 파일과 런타임은 이 컴퓨터에 유지됩니다.",
