+++ apps/web/src/lib/locales/zh-CN.json
@@ -5565,6 +5565,34 @@
     "exportLogsAction": "导出",
     "exportingLogs": "导出中...",
     "connectorTitle": "这台电脑",
+    "clipperTitle": "Shadow Clipper",
+    "clipperDescription": "连接浏览器插件、资料库和本机助手。",
+    "clipperService": "本机连接",
+    "clipperReady": "已启动",
+    "clipperStopped": "未启动",
+    "clipperBrowser": "浏览器插件",
+    "clipperBrowserCount": "已连接 {{count}} 个",
+    "clipperLibrary": "本地资料库",
+    "clipperFileCount": "{{count}} 个文件",
+    "clipperLastSync": "最近同步",
+    "clipperNeverSynced": "尚未同步",
+    "clipperInstallTitle": "准备浏览器插件",
+    "clipperInstallDescription": "桌面端会找到本地插件并打开 Chrome 扩展页。按照 Chrome 提示确认一次“加载已解压的扩展程序”即可。",
+    "clipperPrepareInstall": "准备安装",
+    "clipperConnectTitle": "连接插件",
+    "clipperConnectDescription": "通过“准备安装”加载的插件会自动使用下面的地址和令牌；已有插件也可以手动填写。",
+    "clipperStart": "启动连接",
+    "clipperStop": "停止连接",
+    "clipperAddress": "连接地址",
+    "clipperToken": "连接令牌",
+    "clipperStartFirst": "请先启动连接",
+    "clipperLoginTitle": "同步社区登录",
+    "clipperLoginDescription": "插件连接后，桌面端会自动同步当前社区登录；也可以在这里重新同步。授权一次有效并只交给本机已连接的插件。",
+    "clipperSyncLogin": "同步当前登录",
+    "clipperConnectionStarted": "本机连接已启动。",
+    "clipperConnectionStopped": "本机连接已停止。",
+    "clipperInstallPrepared": "已打开插件目录和 Chrome 扩展页。请选择“加载已解压的扩展程序”。",
+    "clipperLoginQueued": "登录同步已发送，插件会在下一次连接时完成。",
     "connectorWorkspaceOnlineTitle": "这台电脑已接入虾豆",
     "connectorWorkspaceOfflineTitle": "把这台电脑接入虾豆",
     "connectorWorkspaceDesc": "把本地编程工具变成社区 Buddy。虾豆负责把对话和任务送到这里，文件和运行时仍留在这台电脑上。",
