+++ apps/web/src/lib/locales/zh-TW.json
@@ -5225,6 +5225,34 @@
     "exportLogsAction": "匯出",
     "exportingLogs": "匯出中...",
     "connectorTitle": "這台電腦",
+    "clipperTitle": "Shadow Clipper",
+    "clipperDescription": "連接瀏覽器擴充功能、資料庫和本機助理。",
+    "clipperService": "本機連線",
+    "clipperReady": "已啟動",
+    "clipperStopped": "未啟動",
+    "clipperBrowser": "瀏覽器擴充功能",
+    "clipperBrowserCount": "已連接 {{count}} 個",
+    "clipperLibrary": "本機資料庫",
+    "clipperFileCount": "{{count}} 個檔案",
+    "clipperLastSync": "最近同步",
+    "clipperNeverSynced": "尚未同步",
+    "clipperInstallTitle": "準備瀏覽器擴充功能",
+    "clipperInstallDescription": "桌面端會找到本機擴充功能並開啟 Chrome 擴充功能頁面。依照 Chrome 提示確認一次「載入未封裝項目」即可。",
+    "clipperPrepareInstall": "準備安裝",
+    "clipperConnectTitle": "連接擴充功能",
+    "clipperConnectDescription": "透過「準備安裝」載入的擴充功能會自動使用下方地址和權杖；既有安裝也可手動填寫。",
+    "clipperStart": "啟動連線",
+    "clipperStop": "停止連線",
+    "clipperAddress": "連線地址",
+    "clipperToken": "連線權杖",
+    "clipperStartFirst": "請先啟動連線",
+    "clipperLoginTitle": "同步社群登入",
+    "clipperLoginDescription": "擴充功能連接後，桌面端會自動同步目前社群登入；也可以在此重新同步。每次授權僅使用一次，且只提供給已連接的本機擴充功能。",
+    "clipperSyncLogin": "同步目前登入",
+    "clipperConnectionStarted": "本機連線已啟動。",
+    "clipperConnectionStopped": "本機連線已停止。",
+    "clipperInstallPrepared": "已開啟擴充功能資料夾和 Chrome 擴充功能頁面。請選擇「載入未封裝項目」。",
+    "clipperLoginQueued": "登入同步已送出，擴充功能會在下次連線時完成。",
     "connectorWorkspaceOnlineTitle": "這台電腦已接入蝦豆",
     "connectorWorkspaceOfflineTitle": "把這台電腦接入蝦豆",
     "connectorWorkspaceDesc": "把本機程式設計工具變成社群 Buddy。蝦豆負責把對話和任務送到這裡，檔案和執行環境仍留在這台電腦上。",
