+++ packages/cli/package.json
@@ -22,11 +22,11 @@
     "test:e2e": "vitest run __tests__/e2e"
   },
   "dependencies": {
+    "@shadowob/connector": "workspace:*",
     "@shadowob/sdk": "workspace:*",
     "commander": "^13.1.0",
     "chalk": "^5.4.1",
-    "jszip": "^3.10.1",
-    "yaml": "2.8.4"
+    "jszip": "^3.10.1"
   },
   "devDependencies": {
     "@types/node": "^22.15.21",
