+++ packages/connector/README.md
@@ -95,6 +95,23 @@ npx @shadowob/connector@latest update --target cc-connect --server-url https://s
 
 Existing model providers, plugins, projects, platforms, and unrelated keys are preserved.
 
+## Shadow Clipper
+
+Connector owns the Local Bridge runtime used by Shadow Clipper, Shadow Desktop, the Shadow CLI,
+and MCP clients. Shadow Desktop is the recommended setup surface under **Settings → Connector →
+Shadow Clipper**. The existing `shadowob local-bridge` commands remain compatible for scripts and
+advanced CLI/MCP access.
+
+For a lightweight Connector-owned check or foreground service:
+
+```bash
+shadowob-connector clipper doctor
+shadowob-connector clipper start
+```
+
+See the [Local Bridge guide](docs/local-bridge.md) for desktop installation behavior, login sync,
+the complete CLI surface, MCP registration, and security boundaries.
+
 When Shadow's official model proxy is enabled on the server, daemon-created
 Buddy setup jobs can include an OpenAI-compatible provider for runtimes that own
 their provider configuration. Codex launched through cc-connect always keeps
