+++ packages/connector/package.json
@@ -31,6 +31,13 @@
       "import": "./dist/runtime-sessions.js",
       "require": "./dist/runtime-sessions.cjs",
       "default": "./dist/runtime-sessions.js"
+    },
+    "./local-bridge": {
+      "types": "./dist/local-bridge.d.ts",
+      "development": "./src/local-bridge.ts",
+      "import": "./dist/local-bridge.js",
+      "require": "./dist/local-bridge.cjs",
+      "default": "./dist/local-bridge.js"
     }
   },
   "files": [
@@ -79,6 +86,7 @@
   "dependencies": {
     "@shadowob/shared": "workspace:*",
     "dotenv": "^17.4.2",
+    "jszip": "^3.10.1",
     "smol-toml": "^1.6.1",
     "yaml": "^2.8.4"
   }
